SeparatedName&Price

+++ dbm/schema.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS public.costs (
     id integer PRIMARY KEY,
     name TEXT NOT NULL,
-    price integer NOT NULL,
+    price DECIMAL NOT NULL,
     createdtime  TIMESTAMPTZ
 );
