+++ domain/domain.go
@@ -1,8 +1,8 @@
 package domain
 
 type Cost struct {
-	ID          int    `db:"id"`
-	Name        string `db:"name"`
-	Price       int    `db:"price"`
-	CreatedTime string `db:"createdtime"`
+	ID          int     `db:"id"`
+	Name        string  `db:"name"`
+	Price       float64 `db:"price"`
+	CreatedTime string  `db:"createdtime"`
 }
